+++ diive/pkgs/analyses/__init__.py
@@ -1,1 +1,2 @@
 import diive.pkgs.analyses.optimumrange
+import diive.pkgs.analyses.gapfinder