+++ diive/pkgs/corrections/offsetcorrection.py
@@ -3,8 +3,8 @@ from pathlib import Path
 import pandas as pd
 from pandas import Series
 
-from diive.common.dfun.frames import aggregated_as_hires
-from diive.common.dfun.frames import resample_df
+import diive.common.dfun.frames as frames
+# from diive.common.dfun.frames import resample_df
 from diive.pkgs.createvar.potentialradiation import potrad_from_latlon
 
 
@@ -24,7 +24,6 @@ def remove_relativehumidity_offset(series: Series,
     """
 
     print(f"Removing RH offset from {series.name} ...")
-
     outname = series.name
     series.name = "input_data"
 
@@ -34,12 +33,18 @@ def remove_relativehumidity_offset(series: Series,
     exceeds_ix = _series_exceeds.index
 
     # Calculate daily mean of values > 100
-    _daily_mean_above_100 = aggregated_as_hires(aggregate_series=_series_exceeds, to_freq='D', to_agg='mean',
-                                                hires_timestamp=series.index)
+    _daily_mean_above_100 = frames.aggregated_as_hires(aggregate_series=_series_exceeds,
+                                                       to_freq='D',
+                                                       to_agg='mean',
+                                                       hires_timestamp=series.index,
+                                                       interpolate_missing_vals=True)
 
     # Calculate and gap-fill offset values
     _offset = _daily_mean_above_100.sub(100)  # Offset is the difference to 100
+    if _offset.dropna().empty:
+        _offset.loc[:] = 0  # Set offset to zero
     _offset = _offset.interpolate().ffill().bfill()
+    # _offset = _offset.interpolate().ffill().bfill()
     _offset.rename("offset", inplace=True)
 
     # Subtract offset from relative humidity (RH) column (series - offset)
@@ -55,12 +60,12 @@ def remove_relativehumidity_offset(series: Series,
     if saveplot:
         from diive.common.plotting.plotfuncs import quickplot_df
         quickplot_df([series, _series_exceeds, _daily_mean_above_100, _offset, series_corr], subplots=False,
-                     saveplot=saveplot, hline=100, title=f"Removing RH offset from {series.name}")
+                     saveplot=saveplot, hline=100, title=f"Remove RH Offset From {series.name}")
 
     return series_corr
 
 
-def remove_radiation_zero_offset(series: Series,
+def remove_radiation_zero_offset(_series: Series,
                                  lat: float,
                                  lon: float,
                                  show: bool = False,
@@ -76,7 +81,7 @@ def remove_radiation_zero_offset(series: Series,
     nighttime data are set to zero.
 
     Args:
-        series: Data for radiation variable that is corrected
+        _series: Data for radiation variable that is corrected
         lat: Latitude of the location where radiation data were recorded
         lon: Longitude of the location where radiation data were recorded
         show: Show plot
@@ -85,28 +90,32 @@ def remove_radiation_zero_offset(series: Series,
     Corrected series
     """
 
-    print(f"Removing radiation zero-offset from {series.name} ...")
-
-    outname = series.name
-    series.name = "input_data"
+    print(f"Removing radiation zero-offset from {_series.name} ...")
+    outname = _series.name
+    _series.name = "input_data"
 
     # Calculate potential radiation for lower time resolution (30MIN)
-    lower_res, _ = resample_df(df=pd.DataFrame(series), freq_str='30T', agg='mean', mincounts_perc=.9, to_freq='T')
-    potential_radiation = potrad_from_latlon(lat=lat, lon=lon, timestamp_ix=lower_res.index)
-    potential_radiation = potential_radiation.reindex(series.index, method='nearest')  # Reindex to hires timestamp
-    potential_radiation.rename("potential_radiation", inplace=True)
+    lower_res, _ = frames.resample_df(df=pd.DataFrame(_series), freq_str='30T', agg='mean', mincounts_perc=.9,
+                                      to_freq='T')
+    _potential_radiation = potrad_from_latlon(lat=lat, lon=lon, timestamp_ix=lower_res.index)
+    _potential_radiation = _potential_radiation.reindex(_series.index, method='nearest')  # Reindex to hires timestamp
+    _potential_radiation.rename("potential_radiation", inplace=True)
 
     # Detect series nighttime data from potential radiation
-    series_nighttime = series.loc[potential_radiation == 0]
+    series_nighttime = _series.loc[_potential_radiation == 0]
     nighttime_ix = series_nighttime.index
 
     # Calculate offset as the daily nighttime mean
-    offset = aggregated_as_hires(aggregate_series=series_nighttime, to_freq='D', to_agg='mean',
-                                 hires_timestamp=series.index)
+    _offset = frames.aggregated_as_hires(aggregate_series=series_nighttime,
+                                         to_freq='D',
+                                         to_agg='mean',
+                                         hires_timestamp=_series.index,
+                                         interpolate_missing_vals=True)
 
     # Gap-fill offset values
-    offset = offset.interpolate().ffill().bfill()
-    offset.rename("offset", inplace=True)
+    _offset = _offset.fillna(_offset.median())
+    # offset = offset.interpolate().ffill().bfill()
+    _offset.rename("offset", inplace=True)
 
     # Subtract offset from radiation column (rad_col - offset)
     # Offset examples assuming measured radiation is 120:
@@ -116,11 +125,11 @@ def remove_radiation_zero_offset(series: Series,
     #   In case of positive offset +8:
     #       (measured radiation should be zero but +8 was measured)
     #       120 - (+8) = 112
-    series_corr = series.sub(offset)
-    series_corr.rename("corrected_by_offset", inplace=True)
+    _series_corr = _series.sub(_offset)
+    _series_corr.rename("corrected_by_offset", inplace=True)
 
     # Set nighttime radiation to zero (based on potential radiation)
-    series_corr_settozero = series_corr.copy()
+    series_corr_settozero = _series_corr.copy()
     series_corr_settozero.loc[nighttime_ix] = 0
 
     # # Set still remaining negative values to zero
@@ -131,8 +140,16 @@ def remove_radiation_zero_offset(series: Series,
     # Plot
     if saveplot:
         from diive.common.plotting.plotfuncs import quickplot_df
-        quickplot_df([series, potential_radiation, series_corr, series_corr_settozero], subplots=False,
-                     saveplot=saveplot, title=f"Removing radiation zero-offset from {series.name}")
+        quickplot_df([_series, _potential_radiation, _series_corr,
+                      series_corr_settozero, _offset],
+                     subplots=False,
+                     saveplot=saveplot,
+                     title=f"Removing radiation zero-offset from {_series.name}")
+        # Separate plot showing the offset
+        quickplot_df([_offset],
+                     subplots=False,
+                     saveplot=saveplot,
+                     title=f"Used offset values from removing radiation zero-offset from {_series.name}")
 
     return series_corr_settozero