+++ diive/pkgs/flux/carboncost.py
@@ -2,7 +2,6 @@ import matplotlib.gridspec as gridspec
 import matplotlib.pyplot as plt
 from pandas import DataFrame
 
-from diive.common.dfun.frames import insert_aggregated_in_hires
 from diive.pkgs.flux.criticalheatdays import CriticalHeatDays
 
 
@@ -11,25 +10,38 @@ class CarbonCost:
     def __init__(
             self,
             df: DataFrame,
-            x_col: tuple,
-            nee_col: tuple,
-            gpp_col: tuple,
-            reco_col: tuple,
-            daytime_col: tuple,
-            daytime_threshold: int = 20,
+            x_col: str,
+            nee_col: str,
+            gpp_col: str,
+            reco_col: str,
             set_daytime_if: str = 'Larger Than Threshold',
             chd_usebins: int = 10,
-            chd_bootstrap_runs: int = 10
+            chd_bootstrap_runs: int = 11
     ):
+        """
+
+        Args:
+            df:
+            x_col:
+            nee_col:
+            gpp_col:
+            reco_col:
+            set_daytime_if:
+            chd_usebins:
+            chd_bootstrap_runs: Number of bootstrap runs during detection of
+                critical heat days. Must be an odd number. In case an even
+                number is given +1 is added automatically.
+        """
         self.df = df
         self.x_col = x_col
         self.nee_col = nee_col
         self.gpp_col = gpp_col
         self.reco_col = reco_col
-        self.daytime_col = daytime_col
-        self.daytime_threshold = daytime_threshold
         self.set_daytime_if = set_daytime_if
         self.chd_usebins = chd_usebins
+
+        if chd_bootstrap_runs % 2 == 0:
+            chd_bootstrap_runs += 1
         self.chd_bootstrap_runs = chd_bootstrap_runs
 
         # Results from critical heat days analyses
@@ -61,7 +73,7 @@ class CarbonCost:
 
     @property
     def chd_instance(self):
-        """Return results for optimum range"""
+        """Return results for critical heat days"""
         if not self._chd_instance:
             raise Exception('No instance for CriticalHeatDays found')
         return self._chd_instance
@@ -71,23 +83,22 @@ class CarbonCost:
 
     def _criticalheatdays(self):
         """Run analyses for critical heat days"""
-        # Critical heat days (CHDs)
+        # Critical heat days
         chd = CriticalHeatDays(
-            df=self.df,
+            df=df,
             x_col=self.x_col,
-            flux_col=self.nee_col,
+            nee_col=self.nee_col,
             gpp_col=self.gpp_col,
             reco_col=self.reco_col,
-            daynight_split=self.daytime_col,
-            daytime_threshold=self.daytime_threshold,
-            set_daytime_if=self.set_daytime_if,
+            daynight_split='timestamp',
             usebins=self.chd_usebins,
-            bootstrap_runs=self.chd_bootstrap_runs
+            bootstrap_runs=self.chd_bootstrap_runs,
+            bootstrapping_random_state=None
         )
 
         # Run CHD analyses
-        chd.detect_chd_threshold_from_partitioned()
-        chd.analyze_flux(flux='nee')
+        chd.detect_chd_threshold()
+        chd.analyze_daytime()
         chd.find_nee_optimum_range()
 
         # Provide CHD results to class
@@ -97,11 +108,11 @@ class CarbonCost:
 
         return chd
 
-    def plot_chd_threshold_detection(self, ax):
-        self.chd_instance.plot_chd_detection_from_nee(ax=ax)
+    def plot_chd_threshold_detection(self, ax, highlight_year: int = None):
+        self.chd_instance.plot_chd_detection_from_nee(ax=ax, highlight_year=highlight_year)
 
-    def plot_chd_flux_analysis(self, ax):
-        self.chd_instance.plot_daytime_analysis(ax=ax, flux='nee', highlight_year=2019)
+    def plot_daytime_analysis(self, ax, highlight_year: int = None):
+        self.chd_instance.plot_daytime_analysis(ax=ax)
 
     def plot_rolling_bin_aggregates(self, ax):
         self.chd_instance.plot_rolling_bin_aggregates(ax=ax)
@@ -144,19 +155,21 @@ class CarbonCost:
 if __name__ == '__main__':
     from tests.testdata.loadtestdata import loadtestdata
 
-    df = loadtestdata()
+    # Load data
+    data = loadtestdata()
+    df = data['df']
 
-    # Use data from May to Sep only
-    maysep_filter = (df.index.month >= 5) & (df.index.month <= 9)
-    df = df.loc[maysep_filter].copy()
+    # # Use data from May to Sep only
+    # maysep_filter = (df.index.month >= 5) & (df.index.month <= 9)
+    # df = df.loc[maysep_filter].copy()
 
     # Settings
-    x_col = ('VPD_f', '-no-units-')
-    nee_col = ('NEE_CUT_f', '-no-units-')
-    gpp_col = ('GPP_DT_CUT', '-no-units-')
-    reco_col = ('Reco_DT_CUT', '-no-units-')
-    daytime_col = ('Rg_f', '-no-units-')
-    ta_col = ('Tair_f', '-no-units-')
+    x_col = 'VPD_f'
+    nee_col = 'NEE_CUT_f'
+    gpp_col = 'GPP_DT_CUT'
+    reco_col = 'Reco_DT_CUT'
+    daytime_col = 'Rg_f'
+    ta_col = 'Tair_f'
 
     cc = CarbonCost(
         df=df,
@@ -164,33 +177,27 @@ if __name__ == '__main__':
         nee_col=nee_col,
         gpp_col=gpp_col,
         reco_col=reco_col,
-        daytime_col=daytime_col,
-        daytime_threshold=50,
-        set_daytime_if='Larger Than Threshold',
         chd_usebins=0,
-        chd_bootstrap_runs=5
+        chd_bootstrap_runs=3
     )
 
     cc.run()
 
     # Plots
-    fig = plt.figure(figsize=(9, 16))
-    gs = gridspec.GridSpec(2, 1)  # rows, cols
+    fig = plt.figure(figsize=(16, 9))
+    gs = gridspec.GridSpec(1, 2)  # rows, cols
     # gs.update(wspace=.2, hspace=0, left=.1, right=.9, top=.9, bottom=.1)
     ax1 = fig.add_subplot(gs[0, 0])
-    ax2 = fig.add_subplot(gs[1, 0])
+    ax2 = fig.add_subplot(gs[0, 1])
     # ax3 = fig.add_subplot(gs[2, 0])
     # ax4 = fig.add_subplot(gs[3, 0])
     # ax5 = fig.add_subplot(gs[4, 0])
-    cc.plot_chd_threshold_detection(ax=ax1)
-    cc.plot_chd_flux_analysis(ax=ax2)
+    cc.plot_chd_threshold_detection(ax=ax1, highlight_year=2019)
+    cc.plot_daytime_analysis(ax=ax2, highlight_year=2019)
     # cc.plot_rolling_bin_aggregates(ax=ax3)
     # cc.plot_bin_aggregates(ax=ax4)
     # cc.plot_vals_in_optimum_range(ax=ax5)
     fig.show()
-
-
-
 
     # # Insert aggregated values in high-res dataframe
     # _df, agg_col = insert_aggregated_in_hires(df=df, col=x_col, to_freq='D', to_agg='max')