+++ tests/testdata/loadtestdata.py
@@ -4,11 +4,15 @@ from diive.common.io.filereader import ReadFileType
 
 
 def loadtestdata() -> DataFrame:
+    # configfilepath = r'L:\Dropbox\luhk_work\20 - CODING\21 - DIIVE\diive\diive\configs\filetypes\DIIVE_CSV_30MIN.yml'
+    # filetypeconfig = ConfigFileReader(configfilepath=configfilepath).read()
+    filetype = 'CSV_TS-FULL-MIDDLE_30MIN'
     loaddatafile = ReadFileType(
-        filetype='DIIVE_CSV_30MIN',
-        filepath=r'L:\Dropbox\luhk_work\20 - CODING\21 - DIIVE\diive\tests\testdata\testfile_ch-dav_2016-2020.diive.csv')
-    df = loaddatafile._readfile()
-    return df
+        filetype=filetype,
+        # filetypeconfig=filetypeconfig,
+        filepath=r'L:\Dropbox\luhk_work\20 - CODING\21 - DIIVE\diive\tests\testdata\testfile_ch-dav_2016-2020_mayToSep.csv')
+    data = loaddatafile._readfile()
+    return data
 
 
 if __name__ == '__main__':