+++ CHANGELOG.md
@@ -2,24 +2,58 @@
 
 ![DIIVE](images/logo_diive1_256px.png)
 
-## v0.30.0 | XX Jan 2022
+## v0.30.0 | 15 Feb 2022
 
 ### Starting diive library
+
+#### **GENERAL**
 The `diive` library contains packages and modules that aim to facilitate working
 with time series data, in particular ecosystem data.
 
-#### **PACKAGES**
-- Added `pkgs > analyses`: general analyses
-- Added `pkgs > flux`: flux-specific analyses
+Previous versions of `diive` included a GUI. The GUI component will from now on
+be developed separately as `diive-gui`, which makes use of the `diive` library.
 
-#### **MODULES**
+Previous versions of `diive` (up to v0.22.0) can be found in the separate repo 
+[diive-legacy](https://gitlab.ethz.ch/holukas/diive-legacy).
+
+This initial version of the `diive` library contains several first versions of 
+packages that will be extended with the next versions. 
+
+Notable introduction in this version is the package `echires` for working with
+high-resolution eddy covariance data. This package contains the module `fluxdetectionlimit`,
+which allows the calculation of the flux detection limit following Langford et al. (2015).
+
+#### **NEW PACKAGES**
+- Added `common`: Common functionality, e.g. reading data files
+- Added `pkgs > analyses`: General analyses
+- Added `pkgs > corrections`: Calculate corrections for existing variables
+- Added `pkgs > createflag`: Create flag variables, e.g. for quality checks
+- Added `pkgs > createvar`: Calculate new variables, e.g. potential radiation
+- Added `pkgs > echires`: Calculations for eddy covariance high-resolution data, e.g. 20Hz data
+- Added `pkgs > flux`: Flux-specific analyses
+- Added `pkgs > gapfilling`: Gap-filling routines
+- Added `pkgs > outlierdetection`: Outlier detection
+- Added `pkgs > qaqc`: Quality screening for timeseries variables
+
+#### **NEW MODULES**
 - Added `optimumrange` in `pkgs > analyses`
+- Added `gapfinder` in `pkgs > analyses`
+- Added `offsetcorrection` in `pkgs > corrections`
+- Added `setto_threshold` in `pkgs > corrections`
+- Added `outsiderange` in `pkgs > createflag`
+- Added `potentialradiation` in `pkgs > createvar`
+- Added `fluxdetectionlimit` in `pkgs > echires`
+- Added `carboncost` in `pkgs > flux`
 - Added `criticalheatdays` in `pkgs > flux`
-- Added `carboncost` in `pkgs > flux` !!!TODO!!!
+- Added `interpolate` in `pkgs > gapfilling`
+- Added `hampel` in `pkgs > outlierdetection`
+- Added `meteoscreening` in `pkgs > qaqc`
 
-#### **ADDITIONS**
-- XXX
+#### **CHANGES & ADDITIONS**
+- None
 
 #### **BUGFIXES**
-- XXX
+- None
 
+#### **REFERENCES**
+Langford, B., Acton, W., Ammann, C., Valach, A., & Nemitz, E. (2015). Eddy-covariance data with low signal-to-noise ratio: Time-lag determination, uncertainties and limit of detection. Atmospheric Measurement Techniques, 8(10), 4197–4213. https://doi.org/10.5194/amt-8-4197-2015